internal: Move types to separate file (also resolves circular dependency in tsx)
